requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiofiles>=23.2.1
